% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{get_annotation}
\alias{get_annotation}
\title{get gene annotations from a gtf file}
\usage{
get_annotation(
  path_gtf,
  skip = 5,
  gelement = "transcript",
  coding = "protein_coding",
  filter_reg_chr = T
)
}
\arguments{
\item{path_gtf}{path to gtf file}

\item{skip}{number of lines to skip in the gtf file until the first gene annotation.}

\item{gelement}{denotes which gene element are filtered. Available options: "transcript", "exon", "intron"}

\item{coding}{denotes which genes are filtered based on coding information. Available options: "protein_coding",}

\item{filter_reg_chr}{logical. If \code{T}, only genes are included which are on regular chromosomes.}
}
\value{
list containing two slots of full annotation filtered for \code{gelement} and \code{coding} aswell as annotations collpased per gene into gene loci.
\item{annotation}{annotation filtered for \code{gelement} and \code{coding}}
\item{gelement_coding}{annotation filtered for \code{gelement} and \code{coding} and collpased per gene into gene loci}
}
\description{
function to read in gene annotation from a gtf file and filters these annotation based on gene element, coding information and chromosomes.
}
\examples{
anno <- get_annotation(gtf)
}
