% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{give_activity}
\alias{give_activity}
\title{Aggregate peak counts falling onto the same gene to gene activities}
\usage{
give_activity(gene_peaks, peak_matrix)
}
\arguments{
\item{gene_peaks}{data.frame of peak annotations. Output of \code{peak_on_gene} with peak annotated to the genes they fall onto.}

\item{peak_matrix}{Peak-cell count matrix. With rows representing peaks and columns representing cells.}
}
\value{
sparse matrix of aggregated activity counts with rows as genes and columns as cells.
}
\description{
This function takes the output of \code{peaks_on_gene}, run with TSSmode = \code{F} and the peak-cell count matrix as input and aggregates peaks falling onto the same genes to a gene-activity count matrix.
}
\examples{
activity <- give_activity(peak_genes, merged_atac_filt)
}
