% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{give_combined_peaks}
\alias{give_combined_peaks}
\title{Get combined overlapping peak set.}
\usage{
give_combined_peaks(atac_layer, filter_reg_chr = T)
}
\arguments{
\item{atac_layer}{list of peak-cell count matrix of the different samples, with rows as peaks and cells as columns.}

\item{filter_reg_chr}{logical. If \code{T}, only genes are included which are on regular chromosomes..}
}
\value{
GenomicRanges object of combined peaks of different samples of atac_layer list.
}
\description{
This function takes the peaks which have been called on different samples, and thus could represent overlapping peaks with different start and or end positions to provide an overlapping peak set.
}
\examples{
combined.peaks <- give_combined_peaks(atac_layer)
}
