% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge peak-cell count matrices.}
\usage{
merge_samples(list.samples)
}
\arguments{
\item{list.samples}{list with a slot for every sample's peak-cell count matrix}
}
\value{
merged peak-cell count matrix
}
\description{
This function takes the peak-count matrices for different samples, stored in a list with a slot for every sample and merges these matrices.
}
\examples{
atac_layer_merged <- merge_samples(atac_layer)
}
