% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{merged_atac}
\alias{merged_atac}
\title{Replace per sample peaks with union of overlapping peaks and merge to common peak-cell matrix}
\usage{
merged_atac(combined.peaks, atac_layer, insert_run1 = NULL)
}
\arguments{
\item{combined.peaks}{GenomicRanges object of combined peaks. Output of \code{give_combined_peaks} function.}

\item{atac_layer}{list of peak-cell count matrices of the different samples, with rows as peaks and cells as columns.}

\item{insert_run1}{output of function with insert_run1 = \code{NULL}, i.e. vector of collapsed peaks across samples, for which per sample peaks have been replaced by the union of overlapping peaks.}
}
\value{
If \code{insert_run1} = \code{NULL}: vector of collapsed peaks across samples, for which per sample peaks have been replaced by the union of common overlapping peaks. If \code{insert_run1} = output of run with \code{insert_run1} = \code{NULL}, merged peak-cell sparse matrix across different samples with common overlapping peak set.
}
\description{
This function replaces peaks which have been called on different samples, representing overlapping peaks with different start and or end positions, with provided overlapping peaks representing the union of overlapping peaks e.g. output of combined.peaks-function. The harmonized common peak-sets facilitates merging of peak-cell matrices of different samples to a common peak-cell matrix.
}
\examples{
overlap_peaks <- merged_atac(combined.peaks = combined.peaks, atac_layer = atac_layer)
overlap_peaks <- merged_atac(combined.peaks = combined.peaks, atac_layer = atac_layer, insert_run1=overlap_peaks)
}
