% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{too_large}
\alias{too_large}
\title{filter out genes spanning a large range.}
\usage{
too_large(annotations = NULL, gene_element = NULL, tolarge = 5e+05)
}
\arguments{
\item{annotations}{list. Output from \code{get_annotation}.}

\item{gene_element}{data.frame of gene annotations. Either second slot of \code{get_annotation} or similar annotation with second column representing start and third column representing end of a peak.}

\item{tolarge}{number. Indicating maximal range of genes to be included.}
}
\value{
data.frame of filtered annotations.
}
\description{
this function filters out genes which span a region larger than a given threshold.
}
\examples{
anno_prolong <- too_large(gene_element = anno[[2]])
}
