% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{get_annotation}
\alias{get_annotation}
\title{get gene annotations from a gtf file}
\usage{
get_annotation(
  path_gtf,
  skip = 5,
  coding = "protein_coding",
  filter_reg_chr = T,
  TSL = F,
  TSLfilt = NULL,
  filter_transcript_biotype = F
)
}
\arguments{
\item{path_gtf}{path to gtf file}

\item{skip}{number of lines to skip in the gtf file until the first gene annotation.}

\item{coding}{denotes which genes are filtered based on gene biotype and/or transcript biotype information. Available options: "protein_coding", c("processed_transcript", "protein_coding"), etc.}

\item{filter_reg_chr}{logical. If \code{T}, only genes are included which are on regular chromosomes.}

\item{TSL}{logical. If \code{T}, retrieve transcript support level(TSL) for given transcripts.}

\item{TSLfilt}{number. If not \code{NULL} but a number between 1-5 (or a vector containing this values), gene_element is filtered to only contain TSL \%in\% \code{TSLfilt}.}

\item{filter_transcript_biotype}{logical. If \code{T}, filtering based on transcript biotype.}
}
\value{
list containing two slots of annotation filtered for transcripts and \code{coding} as well as annotation collased per gene into gene loci.
\item{annotation}{annotation filtered for transcripts, gene_biotype = \code{coding} and/or transcript_biotype = \code{coding}}
\item{gelement_coding}{annotation filtered for transcripts, gene_biotype = \code{coding} and/or transcript_biotype = \code{coding} and optional TSL. In addition, transcipts of the same gene are collpased per gene into gene loci}
}
\description{
function to read in gene annotation from a gtf file and filters these annotation based on gene element, coding information, hromosomes and transript support level.
}
\examples{
anno <- get_annotation(gtf)
anno_test <- get_annotation(gtf, TSL=T)
anno_test <- get_annotation(gtf, TSL=T, TSLfilt=1)
anno_test <- get_annotation(gtf, TSL=T, TSLfilt=c(1,2))
}
