% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{peaks_closest_gene}
\alias{peaks_closest_gene}
\title{Annotates peaks to closest gene.}
\usage{
peaks_closest_gene(peaks, annotations = NULL, gene_element = NULL)
}
\arguments{
\item{annotations}{list. Output from get \code{get_annotation}.}

\item{gene_element}{data.frame of gene annotations. Either second slot of \code{get_annotation}, output of \code{too_large}, \code{prolong_upstream} or similar annotation with second column representing start and third column representing end of a peak.}

\item{gene_peaks}{data.frame of peak annotations. Output of \code{peak_on_gene} with peak annotated to the genes they fall onto.}
}
\value{
data.frame with a row for each peak, containing chromosome information, start- and end position and the closest proximal gene and if the closest gene is not proximal the closest distal gene as well. Distances to closest (proximal) gene is included. And given input is the output of \code{peak_on_gene}, if peak is falling on gene, this gene represents closest gene.
}
\description{
This function takes the output of \code{peak_on_gene} and annotates peaks which do not fall onto genes to its closest proximal as well as closest distal genes(if closest gene is not proximal). The distances to closest proximal and distal genes is also calculated.
}
\examples{
closest_gene <- peaks_closest_gene(peak_genes,anno_prolong)
}
