% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{peaks_closest_gene}
\alias{peaks_closest_gene}
\title{Annotates peaks to closest gene.}
\usage{
peaks_closest_gene(
  peaks,
  annotations = NULL,
  gene_element = NULL,
  TSSmode = T,
  fast = F
)
}
\arguments{
\item{peaks}{data.frame of peak annotations. Output of \code{peaks_on_gene} with peak annotated to the genes they fall onto.}

\item{annotations}{list. Output from get \code{get_annotation}.}

\item{gene_element}{data.frame of gene annotations. Either second slot of \code{get_annotation}, output of \code{too_large}, \code{prolong_upstream} or similar annotation with second column representing start and third column representing end of a peak.}

\item{TSSmode}{logical. If \code{T} specify if \code{peaks_on_gene} function has been run in TSSmode. ### can be optimized to infer this automatic.}
}
\value{
data.frame with a row for each peak, and columns of peaks_on_gene function and in addition ,closest downstream gene, closest gene in general, distance to gene edge of closest downstream gene and closest gene in general.
}
\description{
This function takes the output of \code{peaks_on_gene} and annotates peaks which do not fall onto genes to its closest downstream as well as closest gene in general(if closest gene is not downstream of the peak). The distances to closest (downstream) gene of each peak is also calculated and returned.
}
\examples{
closest_gene <- peaks_closest_gene(peak_genes,anno)
}
