% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.Peak.annotation.R
\name{peaks_on_gene}
\alias{peaks_on_gene}
\title{Annotate peaks which fall onto genes.}
\usage{
peaks_on_gene(
  peak_features,
  annotations = NULL,
  gene_element = NULL,
  split = "[-:]",
  TSSmode = F
)
}
\arguments{
\item{peak_features}{character vector, e.g. rownames of peak-cell matrix, with chromosome information as well as start and end position separated e.g. by "-" and ":".}

\item{annotations}{list. Output from get \code{get_annotation}.}

\item{gene_element}{data.frame of gene annotations. Either second slot of \code{get_annotation}, output of \code{too_large}, \code{prolong_upstream} or similar annotation with second column representing start and third column representing end of a peak.}

\item{split}{character vector(or object which can be coerced to such), used for splitting each element of \code{peak_features} into a vector with 3 elements containing chromosome information, start= and end position of the peak.}

\item{TSSmode}{logical. If \code{T},assigns whether annotated gene overlaps a TSS and the range the peak spans before TSS, i.e. distance peak-start to TSS (+strand) or peak-end to TSS (-strand)}
}
\value{
data.frame with a row for each peak, containing chromosome information, start- and end position as well as the gene on which the peak is falling or "nomatch" if peak is falling on no gene.
}
\description{
This function, annotates given peaks with the gene(s) they fall onto. Input is as character vector with chromosome, start- end end position. if the peak falls onto the spanning ranges of two genes, gene names are collapse with "|".
}
\examples{
peak_genes <- peaks_on_gene(peak_features = rownames(merged_atac_filt), gene_element = anno_prolong)
}
